Dex patching logic to represent inlinable values in code
Summary:
For all transformation opportunities identified in T189452328, replaced the invoke instruction to the Android Resource API with the appropriate const or const-string opcodes (depending on the type of resource that is being asked for) and moves as necessary to maintain program semantics. Utilized the CFGMutation class.

Also, implemented metrics to report how many resources have been inlined as it occurs

Reviewed By: wsanville

Differential Revision: D59612888

fbshipit-source-id: 84a7b568ac95af4ad5bde4fa483c09f37af1349e