Build cache (#886)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/886

The transition to GA made some mistakes.

The context data needs to be evaluated as an expression. The `$` was missed, we always used one static literal key.

The temporary file was used because of CircleCI restrictions on expressions in the cache key, but is the suggested and supported solution because the file hash is computed lazily.
In GA, it seems the hash is computed eagerly, *before* the file is actually created. On the other hand, the cache key supports general expression evaluation, so we can move those elements directly to the key.

The `github.sha` is not the most up-to-date commit in a PR. However, there does not seem to be one simple best key. So leave that as-is for now and hope that it's good enough or fallback works.

No `restore-keys` have been set to find a less precise cache. Add some.

Differential Revision: D60048474