Do not remove new-instance for classes with dervied finalize
Summary:
Adds logic to the method to get classes with overridden finalize methods such that it gets all finalize methods that are overridden even in parent classes.

Interestingly enough, the DCE pass tests that I added in this diff pass without this change.
The change in the Escaping pass fails on master without this change. Even more interesting, all the tests fail once I add the derived class. Not sure if this is an existing bug that was uncaught or if I did something wrong while adding the tests.
 {F1768967316}

Reviewed By: NTillmann

Differential Revision: D60130999

fbshipit-source-id: e65c922f43ad8d0445422f2dd4f21ed170a37e54