Refactor actions (#894)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/894

As composite actions are not tracked well in the UI, and Github seems not interested in changing that with open issues for many years now, refactor a bit. It means a bit more duplication, but should improve tracking for which steps are expensive (and why rebuilding with build cache still takes a long time).

Differential Revision: D60525075