Fibonacci
Summary:
This adds the option to create auxiliary classes (one per dex), with a static field, clinit, and a function that computes Fibonacci.
If enabled, instead of simple NOP instruction, the nopper then inserts invocations to compute some Fibonacci numbers, and store them in the static field.
This adds somewhat more "realistic" overhead of the injected code.
The option is off by default, so this is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D60621635

fbshipit-source-id: 660e92ccf64c3947ad342341fb66a650004e0ed5