static state is evil
Summary:
This...
```
namespace {

std::atomic<std::unordered_set<DexMethodRef*>*> kotlin_null_assertions{nullptr};
const std::unordered_set<DexMethodRef*>& get_kotlin_null_assertions() { .. }
```

was a bad design: The set of methods is only ever captured once on first access, a somewhat random time. And this is bad because *what should be in the list changes over time*. In particular, the confusingly named `RemoveNullcheckStringArgPass` *creates* new null check methods, which then get matched by the assertion finding code.

{gif:000bwcxo}

This wasn't a problem in the past because the first time we ran the constant-propagation analysis happened to be *after* that pass ran, but this constraint was not encoded or checked anywhere, and again, this whole static thing is just bad design.

Fixing this here by passing through explicitly computed state where appropriate, and leaving it as absent where the whole Kotlin nullness check business doesn't really matter, as we are just doing basic constant-propagation analysis for other purposes.

In practice, this is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D60461017

fbshipit-source-id: eea0adf11a34742c78aa57734949f11f70d56ac5