stable_partitions option
Summary:
For non-coldstart dexes, Redex used to have two possible modes: 1) just emit remaining classes in order, filling up dexes along the way, or 2) minimize cross-dex-references and stuff dexes as much as possible.
Both approaches can result in significant size fluctuations when 1) the app is at the brink of needing another dex, and 2) due to the greedy algorithm that computes a local optimum.

This diff introduces a new option, that allows the partition the classes across a configurable number of fixed dexes. By choosing a relatively large but fixed number, we can avoid the size fluctuations that come from adding or removing a dex.

This new option is off by default, so this is a behavior-preserving diff.

Reviewed By: agampe

Differential Revision: D61037721

fbshipit-source-id: 01085cdf1f7fc9839d7bf7988933392a45208231