Use class frequencies in InterDexReshuffleImpl
Summary:
Use class frequencies in InterDexReshuffleImpl. Added a config `interaction_frequency_threshold` that can be toggled for the APK A/B experiments. For each class, check that it appears in less than or equal to `interaction_frequency_threshold` percent of the time in all interactions. If it's within the threshold, add it to the list of reshuffle-able classes.

Previously, this implementation only considered the classes in ColdStart, which caused performance regressions since the classes might be vital to other interactions we track. Now, we ensure that the classes are not vital to any interactions we track, and we also consider more classes than just the ones in ColdStart.

Differential Revision: D61565988

fbshipit-source-id: b8d91946bdf3aba07228baf6dbdbabbc821aa0f7