Add infrastructure to enable source block testing with a default value
Summary:
Adds some basic piping and infra to allow developers to set a default value for all source blocks. Can be used to test violations that arise irrespective of input profiling data.

I put the code in the `InsertHelper` class in hopes that if we later implement better violation testing (e.g. fuzzing) we could follow a similar paradigm to this diff. However, after looking at the code I can see an argument for creating a completely separate pass if we pursue fuzzing testing in the future. With that said, this diff's changes could be contained solely in `InsertSourceBlocks.cpp` file. I don't feel strongly either way.

Reviewed By: jimmycFB

Differential Revision: D61743203

fbshipit-source-id: 46fe6ad40d569b595ef4dd99f35e85dad57be117