introduce "throughput" interaction option
Summary:
Such interactions are optionally identified via a regular expression on their name.
If we find this all useful, we could move this information into the interaction configuration itself.

If configured, this introduces new thresholds to exclude code in "throughput" interactions.

The option is off by default, so this is a behavior-preserving change.

Reviewed By: ssj933

Differential Revision: D63007580

fbshipit-source-id: 6ad0aea9bb642597071f7738b6002abe129997ff