Find types whose init methods appear to have been inlined in relaxed fashion
Summary:
The input program may have had this type of trick applied already (by d8, r8). Trudge through the code to figure out places where it could have happened, be it Redex or the input app.

As described in the comments, a new instance's type on which `Ljava/lang/Object;.<init>:()V` is invoked is not  included in the result set of `find_complex_init_inlined_types()`, since I am trying to minimize the amount of places in which Redex optimizations will have to be more cautious and I have not seen in practice this type of invocation on an unresolved type causing problems (just non-Object super class invocation).

Used by subsequent diff.

Reviewed By: thezhangwei

Differential Revision: D62672721

fbshipit-source-id: 92fa32d6ff052142aaac2acdc744b801169d25af