Refactor: Always make changes with CFG mutations
Summary:
This diff refactors how mutations happen, to always go through CFGMutation class rather than changing an instruction directly. The way the manipulation logic worked prior to this diff leaves the CFG in an invalid state before the mutations get flushed, which makes it difficult to build upon this to support another more complicated type of transform that needs to be made.

So, always go through CFGMutation, and make some reusable bits as lambda functions that will get called from another place in a subsequent diff.

Reviewed By: NTillmann

Differential Revision: D63414250

fbshipit-source-id: c7667dd1608305eb3b6d65d80a22352463065445