option to treat exception constructors as strict
Summary:
We also don't want to inline constructors of throwable (exception, error) classes, as they capture the current stack trace in a way that is sensitive to inlining.

This new behaviour is introduced via a new flag that is off by default. So this is in effect a behavior-preserving change.

TODO: Explore alternative approach that changes symbolication/positions.

Reviewed By: ssj933

Differential Revision: D63564560

fbshipit-source-id: f520533191c841ec0125b31a89c99e71fb4aa2e9