support model gen
Summary:
Check all writes are safe, which implies that all reads are safe.

Skip unit tests created by model gen, which are named `newCompleteObjectBuilderForUnitTestsOnly`. Example: P1583506530

Model gen classes end in Spec, for example `BroadcastFlowMnetItemSpec`. The generated classes are `BroadcastFlowMnetItem`, and `BroadcastFlowMnetItem$Builder`.

For `Parcel`, the writes occur in `BroadcastFlowMnetItem.writeToParcel`, and the field of interest we're writing is `mEntryPoint` for this case P1581873732.
This field gets written to from the Builder in the `<init>` P1581866539 and when we read from the Parcel P1581790144. This field is already annotated.
The builder code is P1583506706. The setter `setEntryPoint`, accessors, and and builder field `mEntryPoint` are already annotated. Since everything is already annotated, there's nothing to patch.

For `AutoGenJsonHelper`, two additional `$Serializer` and `$Deserializer` classes get generated too P1621713154 P1621722325. The fields we're setting/getting are already annotated. The builder methods of interest look the same and are annotated P1621729171. The base generated class also has annotations P1621736593. The spec also defines a default value for the annotated field, which generates a class P1621717201 that has the appropriate annotations.

Classes produced from modelgen have the ModelDefinition annotation, which we can check when we run into parcel or autogenjsonhelper methods P1621718660.

Reviewed By: thezhangwei

Differential Revision: D63795900

fbshipit-source-id: e16a9062407a89910d5b72e75dfa973eacdff901