Introducing ClinitOutlinePass
Summary:
Certain methods are "hot" in the sense that we know that they get executed in time-critical situations, and we request that they get compiled by ART via baseline profiles. However, ART will never compile clinits.
So, this new pass this pass outlines all code from "hot" clinits to separate methods. The separate methods inherit the "hot"ness, and thus, unlike clinits in general, *will* get compiled by ART on device.

The new pass is not active by default, so this is a behavior-preserving change.

Reviewed By: agampe

Differential Revision: D64493102

fbshipit-source-id: 85bedd00f62590818edbe720b23b947efbe2a071