Implement serializer support for new ResTable_typeSpec header field
Summary:
This field used to be reserved (0) but has been repurposed in recent versions of Android tooling: https://cs.android.com/android/_/android/platform/frameworks/base/+/4f48ffd43b182f31a199e54a39406cbd3d84abb0

This change has been released in binary form in aapt2 for a while, but has not been merged into any of the public AOSP branches until quite recently (so it was unclear what this field was used for previously).

Make the serializer in Redex spit out the value as indicated, count the number of ResTable_type that are emitted then stamp in the header value.

Reviewed By: NTillmann

Differential Revision: D63565323

fbshipit-source-id: d0e7e27377682c1ce8d52195a1ba00c612a9f382