Changes to "hot method cold entry" violation
Summary:
Addressing the revert and comments from D64256522.

To add additional context:

During a basic block meeting, it came up that a method should never be hot/seen and have the first source block of its entry block be cold.

This violation tries to test that. At this point, I'm not sure if our method level appear percentage is good enough for this metric or if the metric provides anything useful.

At the very least, if there are violations introduced after the initial insert source block pass, those might be indicative of a larger issue present in later passes.

Thoughts and comments are welcome. I'm not super attached to this violation, but it ~might~ help with future investigations.

Reviewed By: agampe

Differential Revision: D64489923

fbshipit-source-id: b7419e4cb833323083c31cab09792968d097f074