add an explicit API for creating synthetic SBs
Summary:
The common way to create a synthetic source block is to copy an existing source block and fill it with the synthetic information. Copying with `make_unique<SourceBlock>` also copies the entire chain of source blocks, which causes chain violations.

The new `clone_as_synthetic` methods follow a similar format as the existing `fill_source_block` methods. This will replace instances where we copy a source block without resetting the chain.

Reviewed By: agampe

Differential Revision: D64812788

fbshipit-source-id: 5d85c534c62e3439fe9ac5e33ee9a405135d25db