bug workaround
Summary: We hit this bug in D65169513. See T206169001 for details. Likely triggered by "how ANDROID_DETECTED [...] gets resolved, [...] that's from [...] work on Proguard assume-values support". This is a very targetted workaround that effectively disables the problematic code transformation (and not much else), until we have a proper fix in place.

Reviewed By: itang00

Differential Revision: D65187841

fbshipit-source-id: 39603ebe8e5d0819f94c767c23ce18798a4de0dd