Refactor ResourcesInliningPass::find_transformations
Summary:
The combination of nesting / continue statements here is a bit clunky, especially since `continue;` will skip over the `analyze_instruction()` call at the end of the loop body.

Change this to a lambda so that return early could be used consistently to tidy it up, and to make sure that `analyze_instruction()` happens.

Reviewed By: ssj933

Differential Revision: D64406902

fbshipit-source-id: 320c2b268a4bba234556676daf30cca5a0573824