When splitting or creating a new block, propagate source blocks
Summary: When we split or create a new block upon instruction insertion we used to not add a source block to the new block. This diff propagates the source block values from the last source block in the previous block to a new source block in the newly split/created block.

Reviewed By: jimmycFB

Differential Revision: D61372731

fbshipit-source-id: 7736807f3da06978907d28a556640c562ac10d02