More bounds checks parsing a method
Summary:
Check the number of instructions to fit into the dex.

NOTE: Adds a QoL check that there is not an extraordinary amount of instructions. The threshold is large so as to be improbable to hit in practice, but helps curb allocation size to improve behavior under fuzzing.

Also move all asserts to `INVALID_DEX` versions.

Reviewed By: ssj933

Differential Revision: D66560544

fbshipit-source-id: f25def9da6ce04957f607084b23034936cde3a08