Pipe preprocessed baseline profiles into redex
Summary:
In order to rewire the baseline profile infra to preprocessor -> redex -> postprocessor, we need to supply the output of the preprocessor as an input into redex.

This diff changes the buck files and also adds the output directory of the preprocessor as an input into redex.

Currently this is unused in redex, but the next diff will be constructing the baseline profile off of the preprocessed base.

Reviewed By: jimmycFB

Differential Revision: D65979598

fbshipit-source-id: b47037895e0864275a6dacedd896cf812fbe9b09