Add TypedefAnnoOptPass
Summary:
Add a pass to optimize `valueOf`.

The files TestIntdef.java and TestStringDef.java represent the Falco enums that will have a default `valueOf` method for debug builds and a `valueOfOpt` method that will be edited by this pass. The `map_encoding` string will be filled with the StringTreeMap encoding.

After editing `valueOfOpt`, set it to `public` and switch out all invokes of `valueOf` with `valueOfOpt`

Reviewed By: thezhangwei

Differential Revision: D65912448

fbshipit-source-id: c36177218356ea6cddc1e5df261e7951407aab2a