Revert D67543448: Multisect successfully blamed "D67543445: [redex][string-int-def] Promote "add TypedefAnnoOptPass to the pass list" to staging and stable" for one build failure
Summary:
This diff reverts D67543448
D67543445: [redex][string-int-def] Promote "add TypedefAnnoOptPass to the pass list" to staging and stable by itang00 causes the following build failure:

Tests affected:
- [automation_milan_installer_target30_rdk_x86](https://www.internalfb.com/intern/test/562949998333098/)

Here's the Multisect link:
https://www.internalfb.com/multisect/18023440
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Differential Revision: D67616531

fbshipit-source-id: 5804e968c707850b17983c1fc055e173796fd345