Fix int field descriptor in ir-assembler-test
Summary:
D67722153 points out that the trailing ; here for the primitive field is wonky, leading to problems parsing the expression. The original intent was to allow an int field to be parsed and set the encoded value, and allowing the value to be provided in hex.

So, make the test pass by putting back the original type and hex value, but remove the erroneous ;

Reviewed By: agampe

Differential Revision: D67904778

fbshipit-source-id: c4d2bfb7c19a50ba95d73d58ccdb82477384a55a