"Modernize" FindZlib (#902)
Summary:
As title. Modern boost/cmake versions are dependent on ZLIB::ZLIB, so export that. Move the decision of static vs shared lib into `FindZlib`.

This is somewhat preparing for a time when cmake 3.24 is available on all our supported targets, in which case the static lib choice is supported (the name of the parameter is taken from there). It should be a direct drop-in then.

Pull Request resolved: https://github.com/facebook/redex/pull/902

Reviewed By: NTillmann, shanenay

Differential Revision: D68047031

fbshipit-source-id: 72be73ff6f1857a01e5a5521b76bf32ad24d4b85