Change deepdata input to an unpacked directory
Summary:
We need to pass the deepdata input from baseline profile into Redex. Redex can't uncompress tar.xz files easily.

If we were passing in the baseline profile deepdata input into redex as a single argument, we could just uncompress it in redex.py. However, because we are passing in the config as a file, this would require us to parse the file, extract the deepdata, then mutate the file to the correct path. This seems bad.

Instead, it seems easiest to change the driver to take in a path to a directory instead of to an archive and to move the unpacking of the archive to the time it is downloaded

Reviewed By: jimmycFB

Differential Revision: D66853108

fbshipit-source-id: 2d29d8d317e9526f9e2951a9da31de873f8ce21a