Revert D66907546: Multisect successfully blamed "D66853108: [pogo][baseline] Change deepdata input to an unpacked directory" for one build failure
Summary:
This diff reverts D66907546
D66853108: [pogo][baseline] Change deepdata input to an unpacked directory by kobywchan causes the following build failure:

Tests affected:
- [automation_messenger_instralpha_art_str_sapienz_release](https://www.internalfb.com/intern/test/562950108186736/)

Here's the Multisect link:
https://www.internalfb.com/multisect/19352425
Here are the tasks that are relevant to this breakage:
T209305527: 50+ sandcastle jobs failing for redex

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: kobywchan

Differential Revision: D68264085

fbshipit-source-id: 0303367d033be87cee6d1ee0d462f69be8a71d2b