Revert D66853108: Multisect successfully blamed "D66853108: [pogo][baseline] Change deepdata input to an unpacked directory" for one build failure
Summary:
This diff reverts D66853108
Depends on D68264085
D66853108: [pogo][baseline] Change deepdata input to an unpacked directory by kobywchan causes the following build failure:

Tests affected:
- [automation_messenger_instralpha_art_str_sapienz_release](https://www.internalfb.com/intern/test/562950108186736/)

Here's the Multisect link:
https://www.internalfb.com/multisect/19352425
Here are the tasks that are relevant to this breakage:
T209305527: 50+ sandcastle jobs failing for redex

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: kobywchan

Differential Revision: D68264096

fbshipit-source-id: 2028b6da3c093ef0661a25820a42b5cc66ca8fbe