When checking inlinability at a particular call-site, use the reduced-cfg when available
Summary:
When inlining, a "reduced" cfg of callee at a particular call-site arises when, given particular constant arguments, constant-propagation and local-dce on the callee can reduce its remaining code size.
Previously, this reduced cfg was ignored for some inlinability checks; the code was just never adapted when the concept of a reduced cfg was introduced in the inliner, leaving some checks overly conservative. This is getting addressed here, allowing for more inlining.

Redex runtime overhead: While we now perform more code traversals (for individual reduced cfgs), this only applies to a fractions of all callees. We also move some of the checking from upfront to being interleaved with the inlining. Overall, I don't a significant change in the inliner's runtime overhead.

Motivation: I looked into this just now, as the recently introduced option of "partial" inlining provides for another way of coming up with "reduced" code, and I was a bit surprised that it was rather inefficient when being applied at the IntraDex level. This addresses this inefficiency, giving us the option to apply "partial" inlining either upfront, or late in the Redex pass order, with similar effects.

Reviewed By: thezhangwei

Differential Revision: D68250010

fbshipit-source-id: 90d0d32cc1ba4c59644b572e938c8eefdc1291a8