Fix source blocks in class merging
Summary:
When class merging passes made synthetic blocks, it used to set all the interactions to hit. This is incorrect because class merging doesn't merge only hot methods.

In a block that invokes a callee, add a source block with the same values as the first source block in the callee. For all other blocks in the generated method, take the maximum values from all the source blocks.

Example in IntraDexClassMergingPass:
before inlining the callees: P1727168153
after inlining the callees: P1727168635

a longer example during ClassMergingPass:
before inlining the callees: P1727132029
after inlining the callees: P1727132641

Reviewed By: NTillmann

Differential Revision: D69252991

fbshipit-source-id: cfde12df7de5d1ac3074e0d28814ab5e89ef7f3a