Fix multiple insertion override on the lambda_anno_map
Summary:
The `m_lambda_anno_map` stores a mapping from the class name prefix to a vector of lambda class fields patched. The map facilitate the patching of chained lambdas of the enclosing scope. The map is of type `InsertOnlyConcurrentMap`. However, I realized that the map can be inserted with the same key multiple times. That means that the subsequent insertion will overwrite the previous one with the same key. This unintended overwrite seems to produce a slightly different version of the map causing the patcher to be nondeterministic.

The solution I propose here is to use a simple `ConcurrentMap` to model the `m_lambda_anno_map`.

Reviewed By: ssj933

Differential Revision: D69833477

fbshipit-source-id: e25185eb9265ee5d5e752b67df1577240cfacfd1