Break up the patcher stats into more granular sections
Summary: The current patcher stats are too general. It's impossible to tell which code path can a stats delta be attributed to. Therefore, I propose to break them up into smaller sections. Each section covers a particular top level entry point in the `run()` method of the patcher. Hopefully this change will shed some lights on which code paths are causing the nondeterminism in the patcher.

Reviewed By: ssj933

Differential Revision: D69889601

fbshipit-source-id: 7e80277756fb0f4204957d2a319162bb73c9a62b