Refactor: avoid repeating DexAnnotationSet construction
Summary:
## Context
We repeat the same logic that converts a single Typedef annotation to a `DexAnnotationSet`, and add it to a DexMember.
This is unnecessary, we should only handle a single Typedef annotation at a time. No need to always pass a set around.

In addition, when patching annotations based on existing field/method annotations, we copied the entire `DexAnnotationSet` from a DexMember including non-Typedef annotations like `Lorg/jetbrains/annotations/NotNull;`. This is technically incorrect.

## The Change

This change adds two helpers to add one Typedef annotation to a DexMember by converting it to a `DexAnnotationSet`. The rest of the patcher logic does not need to repeat the same thing again.

Reviewed By: ssj933

Differential Revision: D70521952

fbshipit-source-id: bc12954f022405458acf480c8b8c8af06afd5223