Better remapping support to allow an overlayable id to be deleted
Summary:
D70932003 assumes overlayable ids are reachable roots, which made the serialization code fairly trivial.

This change improves the conventions so that we could actually delete an ID that is overlayable, and confirm that the various structures shrink accordingly. Test cases do this verification, but I also ran `aapt2 d` against the intermediate forms of the test file as it gets deleted (dump text is put in comments for reference).

Reviewed By: ssj933

Differential Revision: D70987082

fbshipit-source-id: 257bdbc17cdb4dbebee168f3991925ccda1831a8