track leaked methods
Summary:
At it stands right now, Redex often doesn't delete methods that have become unreachable, it intentionally leaks them. That is because some passes might hold on to references to those methods. LSAN can find this, so the motivation for this diff is to hold on to those methods so that they don't appear as leaked to LSAN. We also reports statistics on how many methods are affected.

(Long-term, we need to have some kind of weak-reference concept to allow passes to hold on to methods while still allowing to delete them...)

This diff also follows up on my comment here: https://www.internalfb.com/diff/D35115396?transaction_fbid=639671913794607&whitespace=SHOW_ALL --- unifying the behavior of `delete_method` with `delete_field` by always calling `erase_method`, which happens to be the case everywhere anyway.

Reviewed By: agampe

Differential Revision: D71639229

fbshipit-source-id: 54c0fcd8fd7f87c143cda40f9a1b300cd1a41db9