Move skip_anonymous_classes logic to MerchabilityChecker
Summary:
## Context

The previous way we setup the `skip_anonymous_classes` option is risky. What we did was that we removes anonymous classes from the merging_targets of a Model at the beginning. The issue with that approach is that we passed an incomplete hierarchy to the analysis, which ignores that fact that some merging targets can have sub classes that were cut out of the hierarchy. The consequence then is that we might merge non-leaf classes leaving the anonymous leaf class with the incorrect super class.

## Solution

The safer way to apply `skip_anonymous_classes` is to treat them as non mergeables at a later stage. In this way, we don't lose any information in the hierarchy analysis. The mergeability check ensures that we don't merge any non-leaf classes. We also added a test for the anonymous object case we ran into.

Reviewed By: ssj933

Differential Revision: D72095261

fbshipit-source-id: 15a6531e342fe4e02d4e48574b2096866964e00f