fix method-override graph retrieval/computation
Summary:
The code pretends that it can get a method-override-graph from the CSE shared state. This was never case, and we ended up without a method-override graph when also using CSE was enabled, which in turned let to a suboptimal determination of what methods have no side effects (and can be pruned by LocalDCE).

This is getting fixed here.

Reviewed By: wsanville

Differential Revision: D69328885

fbshipit-source-id: 9fd886c9c9f504a229339e0fea388d475dea3658