bug fix
Summary: When the "finder" encounters a default-block that ends with an "inequality", i.e. le/lt/ge/gt, or an equality comparison with a non-constant, then in most cases it would have effectively ignored the -then-branch of the inequality/non-constant-comparison, and moved up the default-block to the goto-target. This is wrong and a provision to restrict the SwitchEquivFinder is introduct.

Reviewed By: ssj933

Differential Revision: D72121880

fbshipit-source-id: ffccbbfe9ac5baf7ac67e56817a4b81505ad5be7