Fix build errors with cmake 4.0 (#914)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/914

CMake 4 now gives an error when the minimum version is below 3.5.
To work around this, we can use a range in `cmake_minimum_required`. The upper bound correspond to the maximum version of supported policies.

Another change is that `FindBoost.cmake` has been deleted, so we need to use `find_package(Boost [...] CONFIG)` now. Boost produces a config file since boost 1.70, and we require 1.71, so this should be fine.

Differential Revision: D72310898