Populate PatchingCandidates for patch_lambdas
Summary:
Reviving the previous stack where we added PatchingCandidates to collect candidate for patching Typedef annotations. By doing that, we don't apply the patching while walking the scope in parallel. Rather, we apply the patching sequentially after the walk to avoid nondeterminism.

In this diff, we only adopting this pattern for the `patch_lambdas` code path.

Reviewed By: NTillmann

Differential Revision: D70583412

fbshipit-source-id: 41f1f336fe37cc7e88dd880aa7064180fcc45441