Simplify parameter patching flow
Summary:
## Context

We infer and patch method parameter for missing Typedef annotation in different fashions.
Sometimes we patch as soon as we find a parameter candidate.
Sometimes we put the parameter candidates in a vector and patch them a bit later.
The logic also branches based on a parameter being nullptr.

## The Change

Here I'm trying to make the flow more consistent. We reduce the branching on the parameter patching logic. Now all the param patching is hoisted up into `patch_parameters_and_returns`. The helper method is invoked in a few placed to find and apply the patching if needed. In most cases, it applies the patching right in `patch_parameters_and_returns`. The only exception is in `patch_lambdas` where it delays the patching one step up. That can potentially be simplified in a following diff.

Reviewed By: NTillmann

Differential Revision: D72750432

fbshipit-source-id: 0aee8d0d63fc7d0ef66efbea2da213b1c69ec329