Introducing UnorderedBag
Summary:
While rewriting many unordered container uses recently, I noticed a pattern where we would sometimes store values in a non-deterministic order in vectors for (memory + performance) efficiency, with really just an implicit contract that the element order cannot be relied on.

This diff introduces a new data type that make this explicit: The "UnorderedBag". It's directly backed by a vector, so same efficiency, but the order in which values are stored is not exposed, unless one goes through explicit (and thus auditable) `UnorderedIterable` or any of the other `unordered_...` global helper functions.

Reviewed By: ssj933

Differential Revision: D72896922

fbshipit-source-id: ac8fa4afc76e1a1ff0ec08b620964a0f7b45e2d7