Use basedirectory in other rules
Summary:
As the proguard spec says the basedirectory should be applied directly.

We attempt to emulate this when loading library jars in `redex-all` by passing in the value. However, the spec does not explicitly disallow specifying multiple `-basedirectory` directives, and `-include`d files may contain their own directive.

The relative path for library jars will no longer resolve when a `-basedirectory` is given. If this was important, restructuring the proguard rules is the way to go.

Refactor `redex-all` a little bit. Make the core code easier to read. Do not attempt to split PG libraryjars.

Add unit test cases.

Reviewed By: NTillmann

Differential Revision: D73529393

fbshipit-source-id: 368025cdf6525e285df2cf1a795a38576a34f1eb