Make the higher 32 bit of SignedConstantDomain of int same as the sign bit
Summary:
Currently, the higher 32 bit of an int is inconsistently modeled, when bit shifting is involved.

This is to be consistent with SignedConstantDomain initialized from a constant. Otherwise, there may be unintended effects such as:

- An int with only highest bit determined to be 1. Left shift by 1 is inferred to be unequal to -2.

Directing enforcing such rule for and/or/xor is unnecessary due to the nature of 2's complement. They are enforced along the way due cleaner structure of the code, and potentially save us from trouble if long time in the future optimization related to casting is implemented.

Reviewed By: NTillmann

Differential Revision: D74787423

fbshipit-source-id: 64c067cc9975a75fc59367e77f085094ed2d6376