Mark more blocks as leafs in SwitchEquivFinder
Summary: As commented in the code, if there are multiple ways to reach what was previously considered a "non-leaf" there is a conceptual problem in which an "extra load" could get used by a non-leaf. This is not represented currently. Transforms using this functionality could lead to an invalid CFG.

Reviewed By: thezhangwei, NTillmann

Differential Revision: D74854243

fbshipit-source-id: 2bf3493960fe332d97df864f8f0562092a8db88d