Change method leak tracking
Summary:
Do not store the methods that maybe leak. Instead mark them to be ignored by LSAN.

In regular builds we do not do anything with the methods and are only interested in the count. So add an explicit count.

NOTE: Other leak tooling like valgrind, jemalloc etc will see these as leaking again.

Reviewed By: NTillmann

Differential Revision: D74913856

fbshipit-source-id: 16f2a501e702239144c63575eefdbb25997ba4ba