Update VisitorTest test case to not depend on hard coded ids
Summary:
This was a hastily written test case, and it just hard coded ids in the sample app resource table that were overlayable at the time of writing. As items get added to the app, these values could change.

Update the test case to look up the ids from their names instead, which will be more resilient.

Reviewed By: ssj933

Differential Revision: D75098004

fbshipit-source-id: 2be307bc04644fd42bafc90bfb37f5448fa1fff2